move benchmarks to stresstest module
