+++ settings.gradle
@@ -8,4 +8,5 @@ include 'es',
         'app',
         'client',
         'sql-parser',
+        'stresstest',
         'udc'
