--- sql/src/test/resources/essetup/mappings/bench.json
@@ -1,64 +1,0 @@
-{
-    "default": {
-        "properties": {
-            "areaInSqKm": {
-                "type": "float"
-            },
-            "capital": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "continent": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "continentName": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "countryCode": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "countryName": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "north": {
-                "type": "float"
-            },
-            "east": {
-                "type": "float"
-            },
-            "south": {
-                "type": "float"
-            },
-            "west": {
-                "type": "float"
-            },
-            "fipsCode": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "currencyCode": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "languages": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "isoAlpha3": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "isoNumeric": {
-                "type": "string",
-                "index": "not_analyzed"
-            },
-            "population": {
-                "type": "integer"
-            }
-        }
-    }
-}
