--- sql/src/test/resources/essetup/settings/bench.json
@@ -1,6 +1,0 @@
-{
-    "index": {
-        "number_of_shards": 2,
-        "number_of_replicas": 0
-    }
-}
