+++ stresstest/src/test/java/io/crate/benchmark/BulkDeleteBenchmark.java
