+++ stresstest/src/test/java/io/crate/benchmark/BulkInsertBenchmark.java
