+++ stresstest/src/test/java/io/crate/benchmark/GroupByArbitraryBenchmark.java
