+++ stresstest/src/test/java/io/crate/benchmark/GroupingProjectorBenchmark.java
