+++ stresstest/src/test/java/io/crate/benchmark/InsertBenchmark.java
