+++ stresstest/src/test/java/io/crate/benchmark/PartitionedBulkInsertBenchmark.java
