+++ testing/build.gradle
@@ -13,8 +13,6 @@ dependencies {
         exclude group: 'org.hamcrest', module: 'hamcrest-core'
     }
     compile 'org.codehaus.jackson:jackson-mapper-asl:1.9.13'
-    compile 'com.carrotsearch:junit-benchmarks:0.7.2'
-
     compile 'org.apache.commons:commons-lang3:3.3.2'
     compile project(':es').sourceSets.test.runtimeClasspath
 }
